rename control to flow
